Prefetch all dependencies before CLI builds
We occasionally hit an error when building plugins due to the builder
CLI parsing not understanding dependency downloads in the output. To
prevent this from happening, we can prefetch dependencies before running
the build.

To accomodate this prerequisite, the pinniped build, and any other
activities we will need to do in the future before actually building the
CLI, this adds a general pre-build-cli target that the various build-cli
targets can depend on to get things ready.

Closes: #122

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>