Set BUILD_SHA to indicate local changes in build
We capture the BUILD_SHA during build and output that in `tanzu
version`. This is useful to be able to tell what a dev build of the CLI
contains. This can be slightly misleading though if there have been
local uncommitted changes included in the build.

To give a clue as to this, this updates the setting of BUILD_SHA in the
makefile to indicate if the local clone is dirty to show there are
additional changes past the last commit.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>