Shallow clone pinniped repo
We clone the pinniped repo just to check out a specific tag and build
it. History other than the version we want is not needed and ends up
getting deleted after the build. To slightly optimize this process,
which will be more significant over time, just clone the state as we
need to build it.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>