Update version of plugins to use git tag
We will adopt the convention that for a major coordinated release of
the tanzu cli that the plugin versions will sync up with the git tag.
More accurately it will sync up with the version of tanzu cli that the
plugin is built with.

Between any two of these releases, the plugin versions can still adopt
a different suffix in its semver.
This is done via overwriting the version field in the plugin's main.go