Throttle retrieval of plugin repos on init
We spin up a bunch of goroutines during init to find and install all
plugins. Depending on the system, this could result in noticable
resource comsumption.

This adds some throttling to this process by limiting the number of
concurrent operations. We base this on the number of CPU cores on the
system. It may be more accurate to calculate the amount of memory usage
needed per operation, and throttle based on the memory in the system,
but this is a simpler method that should, in most cases, have some
correlation between CPU and memory.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>