Delegate help output to the plugins
Plugins are registered with skeleton subcommands to the root command,
then call out to the plugin to perform its action. This causes an issue
for the built in cobra.Command help output since the registered commands
do not have all the subcommand and flag information in order to generate
the full help text.

This adds a custom help function to loaded plugins. It uses the same
execution mechanism to call out to the plugin to get help details.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>
(cherry picked from commit cab1ab865013d3154fd734d1ec2a0cae16e7e705)