Update plugin usage to make sure tanzu root cmd shown
Plugin help is usually generated by calling out to the plugin. That
means that the help output getting its command name will not be
`tanzu plugin_name` but just `plugin_name`.

To make sure the root `tanzu` command is always shown, this adds a check
to the usage template to look for its presence and add it if needed.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>