Handling command line completion from plugins
Our plugins are separate entities that are only represented in the CLI
through a placeholder cobra.Command that is generated based on provided
PluginDescription metadata. Because of this, the built in cobra.Command
completion mechanism isn't able to get completion information from the
plugins as the local command instance is basically empty.

This adds a mechanism where plugins can control how completion
information is delegated to allow flexibility in how plugins are
written.

For "native" cobra.Commands, there is a hidden "__complete" call that
can be used to parse out the completion. We can default to this so
existing plugins should theoretically just work.

Plugins can also have the option to statically define its completion
nouns through the metadata. This is the simplest option where all
possible completions are just reported in the plugin metadata and then
hardcoded into the completion.

To allow plugins to still be somewhat dynamic, there is also the option
to provide a command to be called in the metadata that can be used to
retrieve the list of nouns from the plugin at runtime.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>