Handle command line completion depth
Code was defaulting to just getting the top level completion. This
meant that something like:

```
tanzu cluster [tab][tab]
```

would correctly get the list of completion options for the cluster
plugin. But doing something like:

```
tanzu cluster credentials [tab][tab]
```

would still suggest those first level cluster completions, which would
be incorrect for the `credentials` subcommand.

This updates the cobra.Command based plugin completion logic to
correctly request the completion information for the portion that is
trying to be completed.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>