Allow command line completion depth for dynamic plugins
Command line completions done for plugins that use the dynamic
generation method had previously only been allowed to return the first
level completion nouns for the plugin. This adds the ability for the
plugin to get the full completion request to be able to determine
whether to return completion information for any subcommands.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>