Migrate management-cluster plugin
This migrates the management cluster plugin from the
tanzu-cli-tkg-plugins repo to here. This is a core plugin that is needed
for normal operations, and needs to be included when we make this repo
open source.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>
(cherry picked from commit 0923c310c3f747a8ac2be54c17f9572c31953e2a)