Address linting issues with management-cluster
The management-cluster plugin was previously hosted in the
tanzu-cli-tkg-plugins repo, but is now being moved to core. The linting
checks in the core repo are more extensive than the ones in the previous
repo. This addresses the errors and warnings that were now identified
with the more strict checks.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>
(cherry picked from commit 4db435d88b3b906c6510fd142d86565392f1cf67)