Use buffered channel for init errors
During initialization, when an error is encountered the logic in the
goroutine will send that error to a channel to pass back to the main
control flow. This was using an unbuffered channel though, so if an
error occurred, it would hang waiting for something to pick it up from
the other end.

This switches to using a buffered channel so any errors will queue up,
then be processed later.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>