Suppress duplicate init error output
The RunE cobra.Command method will print the error by default, but we
also print out the error in the main CLI execute() call. This results in
duplicate output on error.

This sets the flag to tell cobra not to print the error so we can print
it in our common log format.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>