Don't indent global flags in usage template
The command flag usage output is a block of all flags, not something
that is iterated over in the usage template. Current local flag template
is correct with no leading indentation, so local flags output is
formatted fine. But global flags has a leading indent, resulting in the
first line of the global flags being extra indented.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>