Introduce Edition to plugin descriptor.
This commit introduces Edition to the plugin descriptor. In the
near-term, some plugins will have context-aware behavior on the basis of
whether they were built for ce or advanced. Namely, the
management-cluster plugin, which facilitates the kick-start UI.

It is at the plugin author's discretion whether to use this variable and
how to expose it. Most plugin authors will propagate this variable using
ldflags at build time.