Enable godoc comment checking during lint
Comment checking in `golangci-lint` is hidden under default exclusion
rules. This commit adds an inclusion rules to check for comments on
exported symbols and packages.

Missing comments on exported symbols are reported by `golint` and
missing package comments are reported by `stylecheck`. This is because
`golint` has a lot of false positives for package comments. This is why
`min-confidence` is adjusted for `golint` to ignore these false
positives (confidence level of missing package comments is `0.2`). The
`stylecheck` linter is much more accurate for this, but does not provide
linting on exported symbol comments, so we have to use both.