Add ability to pass multiple targets to plugin builder
We've limited the number of platforms we target in our Makefile using
the ENV variable. But the plugin builder is not aware of this make
variable, and loops through all known target platforms unless a specific
one is provided.

This updates the builder and the makefile to pass through our ENVS as
multiple --target arguments to the builder. The builder now handles this
as a list of targets and will loop through only the ones provided so we
don't have to waste cycles building things like Windows 386.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>