Remove remnants of tanzu-cli-tkg-plugins bucket
This was an additional repo that housed the management-cluster plugin.
That plugin has since moved to be part of core, so we no longer need to
fetch plugins from a separate bucket to get it.

This actually appears to cause problems now since the management-cluster
plugin is in two buckets and the CLI does not know which one to use.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>