Add ability to build ARM targets
This adds the ability to build linux_arm64 and darwin_arm64 binaries for
things like RaspberryPi's and M1 Macs.

This does NOT add them as default targets that are built in our normal
builds. It only makes it possible to explicitly perform builds for these
targets. Only for testing purposes at this point.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>