Handle help for plugin subcommands
We need to use a custom help function for plugins since there is just a
placehold cobra.Command registered with the root command based off of
the PluginDescriptor information. To get the full help output, the help
request basically needs to be delegated out to the plugin to complete.

This was only working for the top level plugin command. The custom help
function is not passed any additional detail to know what was requested,
so when the call is made, there's no way to know if it was something
like "tanzu help cluster" or "tanzu help cluster credentials update".

This handles the latter by doing some basic parsing of the full command
line that was run to extract out any arguments that need to be passed
along to the plugin for it to know enough to emit the help for what was
actually being requested.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>