Automate bootstrap cluster provisioning for IPv6
Until now, it was required to use the --use-existing-cluster (-e) flag to
deploy an IPv6 management cluster.

Now, when TKG_IP_FAMILY is set to "ipv6" that flag is no longer required; the
automatically provisioned kind bootstrap cluster will be configured to use IPv6

Caveats:

1. A masquerade rule is required for the bootstrap cluster to reach destinations
off of the host machine (like vcenter):

Docker does this for you for IPv4, but not for IPv6, see:
https://github.com/robbertkl/docker-ipv6nat#nat-on-ipv6-are-you-insane

```
networkname=kind
network inspect $networkname | jq -r '.[0].IPAM.Config[1].Subnet') sudo
ip6tables -t nat -A POSTROUTING -s $cidr ! -o docker0 -j MASQUERADE
```

It is unlikely that we want to edit IP tables rules via the CLI, but in the
future, it may be helpful to error message if this rule does not exist.

2. (Maybe) Configure docker daemon with IPv6 nameserver to resolve vcenter

It may be required to configure Docker to use an IPv6 nameserver that can
resolve your vCenter hostname See:
https://docs.docker.com/engine/install/linux-postinstall/#dns-resolver-found-in-resolvconf-and-containers-cant-use-it

Earlier documentation recommended updating the CoreDNS corefile to address this
need, which was easy to do when using an existing cluster. by configuring the
docker daemons that is no longer required.

Signed-off-by: Christian Ang <angc@vmware.com>
Co-authored-by: Aidan Obley <aobley@vmware.com>
Co-authored-by: David McClure <dmcclure@vmware.com>