Add pod and service CIDR in kind cluster
Cluster and service CIDR are required by kind cluster since the NO_PROXY
parameter on kind and TKG clusters are the same. If the CIDRs being used
in kind cluster is different from TKG cluster, then NO_PROXY will
forward in-cluster traffic to the proxy which would be a problem.

Signed-off-by: Hanlin Shi <shihanlin9@gmail.com>