For AWS, automatically pivot management cluster to only using EC2 instance profile credentials instead of hard-coded credentials file during init and upgrade.
Zeroes-out the secret that is mounted to ~/.aws/config in the pod,
so CAPA uses default credential provider chain.

The deployment's affinity rules are then set such that it
is then pinned to only run on control plane nodes.

As fakes had to be regenerated, and there wasn't a make task for
running counterfeiter, this has now been added, and upgraded to v6,
regenerating all of the fakes.

Signed-off-by: Naadir Jeewa <jeewan@vmware.com>