fix regression on cluster upgrade
A regression was introduced in #120 that caused failures when an invalid
value for options.Namespace is passed in to retriveRegionalClusterConfiguration.

We now get the cluster name from the current context. Then we use the
name to fetch a list of all clusters and use the cluster name to discover
the namespace.

Signed-off-by: Aidan Obley <aobley@vmware.com>