Fix TKG_CUSTOM_IMAGE_REPOSITORY not respected during mc delete
We are seeing issue when we switch the custom image repository during
upgrade or deleting management cluster using TKG_CUSTOM_IMAGE_REPOSITORY
it still tries to use the image repository with which management cluster
was created.

This leads to errors when using different image repository for testing
upgrade and deleting management cluster because it tries to fetch images
from old registry.

This fix takes TKG_CUSTOM_IMAGE_REPOSITORY config variable into account
if specified. If not it uses image repository from management cluster

Signed-off-by: Anuj Chaudhari <anujc@vmware.com>