Allow plugin native completion to propagate ShellCompDirective (#180)
Cobra completion commands pass a directive as the last line of the
output that indicates how the completion script should interpret the
completion values. For example, the value of an argument or flag may be
a file path.

The native completion handler for plugins is currently hard coded to
return ShellCompDirectiveNoFileComp for all sets of completion,
regardless of what value the plugin returned. Now the directive returned
from plugins is propagated to the tanzu completion script. If an error
occurs, ShellCompDirectiveError is returned.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>