Add a script to help manage GitHub labels
To help create a review/approval process for changes to GitHub labels,
introduce a script (based on one from another VMware project[1]) which
uses the GitHub API to create and update labels.

This allows changes to GitHub labels and their descriptions to be
proposed using pull requests. Once merged, a maintainer can use an API
token to apply the change. The history of the script can be used to
understand the reason a GitHub label was introduced.

This does not prevent direct manipulation of GitHub labels by users with
the permissions to do so.

[1]:https://github.com/vmware/vic-tools