Network resource name in vsphere to include path to uniquely identify network name
UI executes getNetworks() API and currently for vSphere, backend is returning the network name and MOID. This is causing issue identifying the unique network if the vSphere has duplicate network names. This PR would fix this issue by passing the network name with full path instead of just network name.

Signed-off-by: PremKumar Kalle <pkalle@vmware.com>