Add gRPC keep-alive time and timeout
By default the gRPC client does not send a keep-alive ping. So, in
case of broken connections, it would take the system TCP timeout
(can be up to 30 minutes) to detect the failure. Adding a client
side keep-alive helps to identify this break earlier. Also, this is
configured so that the server does not complain of too many pings.
Ref.: https://github.com/grpc/grpc-go/blob/master/Documentation/keepalive.md

Signed-off-by: Giri <gelumalai@vmware.com>