Fix Bootstrap cluster reuse issue (#367)
This change removes ClusterResourceSet objects from a management cluster
before deleting the management cluster. This is necessary when using a
tkg workload cluster as bootstrap cluster. In this scenario the
ClusterResourceSet objects cannot be removed once moved to the bootstrap
cluster and this results in future cluster creations failing on this
workload cluster.

Signed-off-by: Nick Tenczar <ntenczar@pivotal.io>