set ceip to false by default for TCE
Before the commit, the current logic made users that did not specify the
ceip flag opt into CEIP. This has been decided as a desired before for
commericial tanzu-editions. However, for Tanzu Community Edition (TCE),
we want this to be set to false by deafult.

This commit uses the BUILD_EDITION (linker flag) to determine if the
edition is "tce" and then set CEIP to false. Otherwise, it will continue
to be the default, which is true.

Signed-off-by: joshrosso <rossoj@vmware.com>