Use custom steps for standalone cluster init
We've had an issue where the deployment UI shows steps during deployment
that mention management clusters instead of standalone clusters. This
was due to the steps the standalone cluster plugin was sending to the
UI. Progress would also fail to update due to a mismatch between the
step reported and the available steps.

This updates the deployment sequence steps for the standalone cluster
init so they only mention standalone clusters and they match the actual
steps taken.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>