TKR Controller: begin TKR discovery without random wait
Before this change, TKR Controller was waiting for a random
duration instead of immediately starting the TKR discovery.

This was happening because r.SyncReleases() call was placed in a select
statement under `case <-ticker.C:`. Ticker is known to drop ticks, i.e.
their delivery is not guaranteed (if the receiver is not ready, for
example).

Additionally, errors when fetching BOM images were not returned
(only logged), which prevented from retrying the failed operation.

Signed-off-by: Ivan Mikushin <imikushin@vmware.com>