Encode build information into the built binaries
This is necessary to trace the exact commit used to build the
binaries used as controllers or CLI commands in a uniform way.

Now, any controllers built from this repo will emit this
version information in one of the first lines printed in the
log.

It is worth noting that before this change, this build
information was encoded and used in a few different places.
This is an attempt to fix it. A single package holds variables
into which version information is injected at build time.

Other variables that used to hold similar information are being
deprecated and are pointing to the one and only true buildinfo
package.

Signed-off-by: Ivan Mikushin <imikushin@vmware.com>