gRPC: Remove context timeout for streaming RPCs (#385)
* gRPC: Remove context timeout for streaming RPCs

Having a timeout in the streaming RPCs causes the streaming request
to exit before getting any response. Hence, removed the context
timeout added by default in the streaming client interceptor.

Signed-off-by: Giri <gelumalai@vmware.com>

* Add a new `pkg/v1/grpc` package for gRPC client utilities

Currently, these utilities are under `pkg/v1/auth/csp` which is not
the ideal location for all gRPC utilities. So, created a new package.
All auth related utilities for gRPC are still in the old package and
are imported in the new one as needed.

Signed-off-by: Giri <gelumalai@vmware.com>

* Add gRPC keep-alive time and timeout

By default the gRPC client does not send a keep-alive ping. So, in
case of broken connections, it would take the system TCP timeout
(can be up to 30 minutes) to detect the failure. Adding a client
side keep-alive helps to identify this break earlier. Also, this is
configured so that the server does not complain of too many pings.
Ref.: https://github.com/grpc/grpc-go/blob/master/Documentation/keepalive.md

Signed-off-by: Giri <gelumalai@vmware.com>