Add markdown linting to set of linters
Following what is done in the community-edition repo, this adds the use
of the mdlint tool to enable linting of our markdown files. This will
help ensure we have good structure and syntax to help avoid rendering
issues.

When errors are encountered with the mdlint linting job, if you aren't
very familiar with markdown it can be kind of confusing understanding
what the issues are and how you are supposed to fix them.

To help give some breadcrumbs to folks in this situation, this adds a
couple places to a link that contains a list of the rules with some
common errors illustrated, along with suggestions for how to correct
them.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>