Fix capability discovery tests due to client-go changes
This commit accounts for changes in FakeDynamicClient in client-go. It
appears the fake dynamic client does not exactly support *List types
(maybe a bug?), so cloudprovider and infraprovider tests are adjusted
just to work with non-list Kinds.

Capability query tests are fixed by using the right scheme because
FakeDynamicClient now requires proper scheme to be registered. The
previous scheme was wrong, but it was being ignored by the fake client.