Add pull request process to CONTRIBUTING docs
Once a pull request is opened and the review process begins, it is not clear what is needed in order to get the pull request merged.

At a high level, we want to try to make the following clear for contributors:
- What is the next step towards getting this PR merged?
- What's the status of that next step?
- Is there anything I can/should do to help make that happen?

In doing so, the process aims to clarify the following:
- Who is responsible for _reviewing_ my pull requests?
- Who is responsible for _approving_ pull requests?
- Who is responsible for _merging_ pull requests?
- What should I do if it seems like my pull request has gotten "lost in the mix"? How should I determine whether any further action is needed on my part?

This is a starting point for documenting a process that we can follow
and then further improve.

Related: https://github.com/vmware-tanzu/tanzu-framework/issues/844