addons/pinniped/post-deploy: support cluster-scoped concierge (#949)
Pinniped v0.6.0 changes Concierge APIs to be cluster-scoped. We want this
post-deploy job to work with v0.12.0, so we need to change the Concierge client
to be cluster-scoped. However, we don't want to immediately update the Concierge
client until the new v0.12.0 package is shipped, so we will do this dynamically
via a command line flag for now. Eventually, after we no longer need to support
any Pinniped versions before v0.6.0, we can hardcode the Concierge client to be
cluster-scoped.

Signed-off-by: Andrew Keesler <akeesler@vmware.com>