cmd/cli/plugin/pinniped-auth: assets should be empty files (#973)
The embedded pinniped CLI was previously an empty file that was
checked into git. We checked it into the repo so that the go:embed
directive always worked. We accidentally checked the actual binaries
into the repo when we did the work to add the 0.12.0 pinniped CLI
alongside the existing 0.4.4 pinniped CLI.

Signed-off-by: Andrew Keesler <akeesler@vmware.com>