Register cleanup for antrea and disable secretgen and vsphere-az1/2
Antrea fails to start after Windows nodes reboot because OVS bridge and
HnsNetwork previously created by Antrea is not cleaned, a cleanup script
is provided in the new version of Antrea, this change registers the
script as a callback on system's shutdown, the cleanup script will be
invoked before Windows shutdown.

secretgen does not support Windows cluster, just disable it.

Disable vsphere-az1/2 on Windows cluster because this is not tested on
Windows clusters.