Update golangci-lint to 1.43.0
This updates the version of golangci-lint we use for code linting to a
version that works with go 1.17. The updated version pulls in various
linter updates, so this also includes fixes for the issues those newer
lint checks were now flagging.

This also drops the use of the gomnd linter for magic number checking.
In general this is a good thing to check, but in nearly all cases where
this was flagging an issue in our code, we would just nolint the error
and keep using the numbers. So rather than enabling it globally and
disabling it locally, this just disables it globally until we decide if
this is something we want to take the effort to address.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>