Update generated data files
There are a few files that are generated and have changes that have not
been merged in to the repo. This is causing problems with some PRs since
we have a `verify` target that is run that makes sure there aren't any
changed files after performing a build.

This includes changes related to newer versions of go, but does not
remove the compatibility for the official 1.16 version we declare. This
shouldn't cause any harm for now, and will make for less issues if/when
we decide to update to go 1.17 in the future.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>