Fix file count error returned by OCIArtifact.Fetch
When the OCI bundle for a CLI plugin contains several files, the OCIArtifact.Fethc()
method bails out.

Since errors.Wrapf returns nil if the err passed in was nil, this would
manifest as a (nil []byte, nil error) return from the Fetch() function,
and the plugin installation would end up writing an empty file to disk,
finally leading to a failed plugin installation when the 'info'
subcommand was executed, but with no obvious error message about why.

To add tests, I extended the OCIArtifact struct to inject the helper
function to construct the file map.