Add management-cluster audit logging
This adds new functionality where an audit log will be create for all
management clusters on create. This will capture all logging activity,
regardless of set log verbosity. This can be used to help troubleshoot
bootstrap and other failures by having everything that happened during
deployment available in a timestamped log.

If the management cluster is upgraded, we will attempt to capture
upgrade logging activity to the log so it is captured in case it is
needed later.

This file is kept around until deleted by the user or until management
cluster deletion. The delete operation will log everything to this audit
log in case of deletion failure. If that fails, the user and/or support
person can review the log file to help troubleshoot. On successful
deletion of the cluster the audit log file will be removed.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>