Remove bold formatting from table output
The current output formatting for table output bolds the header text.
This has been shown to cause issues when running commands under the
`watch` command or in environments where the terminal is not a full tty.

This change removes the bold formatting. Many other command lines that
emit table output, such as kubectl, do not bold the headers. So the
easiest solution for us is to just stop doing so.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>