Fix the CLI plugin manager (#1493)
Commit `4ca1376` changed the signature of a function. Though it
updated the all the previous usage of the said function, there
was another commit `aeea04d` that added a new usage of the same
function. Since `4ca1376` was not rebased on `aeea04d` the build
was broken. Fixing the new usage of the function.

Signed-off-by: Giri <gelumalai@vmware.com>