[WIP]Kickstart UI: Refactor swagger generated API client into external library
Refactors swagger generated API client in the UI out to a standalone Angular library. Where the previous API client was consumed locally at a relative path, it is now imported via the library.

This contains some of the tooling for building the library but some finalization will need to be done.