Fix TKR status not being updated after mgmt cluster upgrade
Add a watch on Cluster and TanzuKubernetesRelease resources,
update TKR status conditions on changes.

Use PatchSet to calculate and apply patches to multiple objects.

Before this change,
`tanzu management-cluster upgrade` would often leave the management
cluster in a state when there would be no compatible TKR or TKR
Compatible status condition wouldn't be updated. This would happen
because the TKR Controller did not watch the Cluster resource
for the management cluster, which is updated as the very last step of
`tanzu management-cluster upgrade`.

Signed-off-by: Ivan Mikushin <imikushin@vmware.com>