Retract v1.4.0-pre-alpha-* tags in go.mod
They are legacy tags created before v0.1.0 was created. Since they are
higher semver-wise than the latest release (v0.15.0 at the time of this
commit), Go proxy thinks that is the highest release of this repo. After
the next release of framework which contains this commit, doing `go list
-m -versions` will not list the retracted versions.

Retracting these versions also means anyone adding an untagged commit of
framework as a dependency in their go.mod will not get a confusing
version that looks like `v1.4.0-pre-alpha-2.0.20220125212408-<SHA>`. It
will have a version that reflects the commit relative to the latest tag.