Add dependabot configuration
This adds configuration for the GitHub Dependabot. With this
configuration, Dependabot will automatically propose updates for
dependencies. We can then choose whether to merge those updates, or if
there is some reason to wait, keep the PR around or close it. At least
this way we are aware of possible updates so we can be less reactionary
towards security and other update needs.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>