CLI: Add pre-download verificatin for HTTP plugin artifacts
Currently, this trusts any local path and the "tanzu-cli-advanced-plugins"
GCP bucket where TMC plugins are stored. Other exceptions can be added
as and when necessary.

Also, switches the default behavior from allow to deny.

This is needed to tighten the security hooks and disallow installation
of plugins from untrusted locations in case a user logs into an
untrusted cluster.

Signed-off-by: Giri <gelumalai@vmware.com>