Remove "trivial" exception from CLA note
We had text in the contributing guide saying that trivial changes could
still be made without signing the CLA. That was inaccurate, and a signed
CLA is required for any kind of change made to the repo.

This updates that text to make this clear and adds a suggestion that
they may file an issue for the maintainers to make the change if they
are not able to sign the CLA and do it themselves.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>