Remove double compile of core from builder
The `builder cli compile` command currently builds the "core" tanzu
binary twice - once with the passed in version number, and once with the
version set to "latest".

It looks like this "latest" version may have been for a purpose that was
never fully realized, and there are no current uses of this extra
binary. The signing process we have explicitly ignores this artifact as
well.

This is a minor savings, but since we do not use or need this "latest"
binary, this updates the builder command to only compile once for the
version requested.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>