Detect go modules for linting
We've had the issue that any time we need to add a new go module to the
repo, the person adding it needs to remember to include it in the
Makefile as an additional location for the `go-lint` and `modules`
target since those operate on a per-module level.

This updates those targets so they automatically iterate through all
detected directories containing a `go.mod` file. That way when someone
adds new modules in the future, they don't need to be as aware of this
linting limitation.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>