Clean up and organize `builder` command code
The builder command has evolved over time as the project has changed.
This cleans up the code and fixes things like defaults for newly
generated plugins so they better match with current requirements.

This also reorganzies some of the code to better separate the
presentation layer (/cmd/cli/plugin*) from the business logic
(pkg/v1/builder). This has the added benefit of making it easier for
clients to consume the code if they need to programmatically perform
some of the operations.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>