Check out current source for linting
It was noticed in another PR that changes to the linter target were not
picked up, missing some issues. On investigation, it was found that this
was a configuration issue in the Lint job. GitHub Actions need to
explicitly check out the code of the PR to pick up the proposed changes,
so even though the PR made changes, they were not reflected in the
results of the job.

This adds the checkout action to the Lint job to make sure it always
runs against the proposed changes.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>