move back to build flag edition specification (#1800)
This commit returns framework to setting the edition based on a build
flag. Due to v0.11.2 cherry-picking many commits, its picked up new
edition detection features that are missing ancillary commits. In
summary, setting an edition is entirely broken.

This reintroduces the edition build flag and still defaults builds to be
tkg, yet allows the TCE team to build a TCE version.

Signed-off-by: joshrosso <rossoj@vmware.com>