Add script to download and install Tanzu Framework CLI (#1705)
* feat: 🎸 Add script to download and install tanzu cli

Currently there is a lengthy way to install tanzu cli in the documentation. Provide an easier option to install tanzu cli by just calling a script which downloads the latest stable release and install the same and update the documentation to reflect the new way of installation.

* refactor: 💡 Resolve comments from PR 1705

Resolve the comments in .gitignore, documentations and copyright related
changes suggested in PR 1705

BREAKING CHANGE: 🧨 N/A

* Correct the usage of fetch and install cli script

The usage guideline in the comments are erroneous. Correct it to point to right usage

Co-authored-by: Navid Shaikh <navids@vmware.com>