Automate release note generation process for Tanzu Framework Release (#1707)
* feat: 🎸 Automate release note generation process

Auto generate release note for every new release. When tag is pushed, automate the release note creation for the new release.

BREAKING CHANGE: 🧨 NA

* fix: 🐛 Release branch for patch should not be main

The base branch from which a tag is cut for patch is not main but the
release-x.y branch. Hence for patch releases, the branch in release-note
generation should point to the release-x.y branch rather than the main
branch.

* fix: 🐛 Re-enable the build to happen when release workflow runs

For speedy testing purpuse, the build job in release workflow was
disabled. Enable it again so that build happens prior to other
activities.