Add v1alpha2 API types
This version of the TKR API is used in a Tanzu distribution
known as TKG Service. It is basically identical to TKR v1alpha1.

We need this to be able to build the TKR v1a2->v1a3 conversion webhook.
The conversion code is also included
(see v1alpha2/tanzukubernetesrelease_conversion.go).

We also include other objects for the sake of completeness.
The v1alpha2 package can now be fully imported from tanzu-framework.

Signed-off-by: Ivan Mikushin <imikushin@vmware.com>